+++ lib/Types.mli
@@ -50,9 +50,12 @@ type error =
   | Parse_error of yaml_error
   | Expansion_limit_exceeded of int
   | Depth_limit_exceeded of int
-  | Plain_error of string
+  | Printer_error of string
   | Document_count_error of string
   | Schema_error of yaml_error
+  | Simplicity_error of yaml_error
+      (** A YAML feature not allowed in simple mode was encountered: an anchor,
+          alias, explicit tag, or (in YAML 1.1 mode) a merge key ([<<]). *)
 
 exception Error of error